api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
Pillow>=10.0.0
pytest>=8.0.0
httpx>=0.27.0
